// services/api-service/src/whales/dto/whale-query.dto.d.ts
export declare class WhaleFeedQueryDto {
    minSize?: string;
    marketId?: string;
    walletAddress?: string;
    page?: number;
    limit?: number;
}
export declare class WhaleTopQueryDto {
    sortBy?: string;
    period?: string;
    limit?: number;
}
//# sourceMappingURL=whale-query.dto.d.ts.map